mac_factory/requirements.txt
litellm>=1.80.0
pyyaml>=6.0
python-dotenv>=1.0.0
